bugfix: properly indent sibling messages
I lost the left inset on sibling messages when I split their logic
out from the unselected message group. This just restores it. I know
that this means unselected and sibling have a lot of code duplication,
but it's only 3 lines and I expect them to drift apart in the future.