perf: prefetch identity and community data to simplify render
This modifies the reply list so that each element is automatically populated
with a pointer to the author and community of a node in addition to the
reply data itself. This ensures that we do not need to do acquire this data
manually within the layout phase of the reply view. The performance increase
is negligible, but the simplification to the render code is worth it.