feat: ensure reply view always has focused message
This commit changes the interaction model for the reply view such
that there is always a message in focus. To summon the contextual
app bar, you tap/click the focused message again. To change focus,
you click/tap on a different message. This eliminates the useless
state in which the ancestry of nothing is highlighted. This also
adds some left margin to unselected messages.