ci/cd: improve flatpak make targets and cleaning
This commit adds a `make fp-shell` target that spawns a shell inside
of the resulting flatpak from `make fp`. This is useful for checking
the sizes of the files that have been inserted.

This also adds some cleaning steps to the flatpak manifest to ensure
that none of the go distribution or cache is included in the final
flatpatk.