fix: stabilise scroll bar height
Cache the maximum visible replies and use that as the hueristic to calculate
scrollbar height. This avoids the scrollbar height changing dynamically as user
scrolls the list. Bar height will grow initially but it will stabilise once the
maximum replies on screen has been encountered.

`numReplies` was just a duplicate of the already cached `c.replyCount`, and was
removed in favour of the latter.

Signed-off-by: Jack Mordaunt <jackmordaunt@gmail.com>