feat: create separate conversation and message filter levels
This commit introduces a new "Conversation" filter that shows sibling
messages in addition to direct descendants and ancestry. The three filter
modes are cycled through by pressing spacebar, clicking the filter button,
or double-tapping a message.

Note: I am not happy with the current UX. I think that often, there's no
difference between the conversation and message filter modes. I think it's
confusing to always have to pass through all three levels on your way to
unfiltered. I'm releasing this version in an attempt to garner ideas and
feedback on how this interaction *should* be implemented.