fix: ensure view is invalidated whenever new nodes arrive
This is an attempt to fix a problem in which the state would be updated
with new nodes, but the UI wouldn't update to display them until the user
interacted in a way that invalidated the window. The problem is hard to
trigger deliberately, but I believe this should handle the relevant
edge case.