fix: ensure back button doesn't return to view id zero
Previously, on android, pressing back from the reply view would take
you back to the view with id 0 as an artifact of the implementation of
RequestViewSwitch and how it pushes to the stack.

By using SetView to select the first view, we prevent a zero from being
pushed onto the view stack.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>