skelsprig: begin refactoring using skel
This commit begins the long process of restructuring sprig
to use primitives offered by the skel Gio framework.

As this is a radical applicationi architecture change,
I'm doing it in a subdirectory, gradually migrating code
from sprig's previous incarnation as it makes sense to
do so.

This commit just demonstrates a work-in-progress settings
page that can read and edit the existing sprig settings
page.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>